backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
motor==2.5.1
yfinance==0.1.63
pandas==1.3.3
pydantic==1.8.2
python-dotenv==0.19.0
openai==1.3.0
requests==2.31.0 